+++ API/src/main/java/kodlamaio/hrms/business/abstracts/JobAdvertiseService.java
@@ -12,7 +12,7 @@ public interface JobAdvertiseService {
     DataResult<List<JobAdvertise>> findAllByActiveTrue();
     ResponseEntity<DataResult<List<JobAdvertise>>> findAllByActiveTrueSorted(int sortId);
     ResponseEntity<DataResult<List<JobAdvertise>>> findAllByActiveTrueAndEmployerUuid(UUID uuid);
-    ResponseEntity<DataResult<List<JobAdvertise>>> findAllByActiveTrueAndCityId(Long id);
+    ResponseEntity<DataResult<List<JobAdvertise>>> findAllByActiveTrueAndCityId(Short id);
     ResponseEntity<DataResult<List<JobAdvertise>>> findAllByActiveTrueAndJobPositionId(Long id);
     ResponseEntity<DataResult<List<JobAdvertise>>> findAllByActiveTrueAndEmployer_CompanyNameContainsIgnoreCase(String companyName);
     ResponseEntity<DataResult<List<JobAdvertise>>> findAllByActiveTruePaged(int pageNumber, int pageSize);
