+++ API/src/main/java/kodlamaio/hrms/repositories/CityDao.java
@@ -3,5 +3,5 @@ package kodlamaio.hrms.repositories;
 import kodlamaio.hrms.model.concretes.City;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CityDao extends JpaRepository<City, Long> {
+public interface CityDao extends JpaRepository<City, Short> {
 }
